+++ emulator_functions.h
@@ -15,8 +15,8 @@
 
 /* Source Instruction Operations */
 
-uint32_t get_code8(Emulator *emu, int index);
-int32_t get_sign_code8(Emulator *emu, int index);
+uint8_t get_code8(Emulator *emu, int index);
+int8_t get_sign_code8(Emulator *emu, int index);
 
 uint32_t get_code32(Emulator *emu, int index);
 int32_t get_sign_code32(Emulator *emu, int index);
